backend/requirements.txt

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.3.0,<2.0.0
langchain==0.1.16
langchain-core==0.1.49
langchain-community==0.0.35
qdrant-client>=1.7.0
motor>=3.7.1
pymongo[gridfs]>=4.6.0
pypdf>=4.0.0
python-docx>=1.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
faiss-cpu==1.7.4